use npm start directly
If your package start has more than one command in it, ploy was chopping it up and trying to run ps with the first command and the rest as arguments to it. Here is an example of my start:

        "start": "couch/prestart; node index.js;",

The above commit calls npm start directly if a pkg.script.start is available. npm is much more stable at invoking the start.

Also I have tried to take into account the windows users.